requirements.txt
streamlit
requests
fastapi
uvicorn
google-genai
python-dotenv
pydantic

